Corrected ms_mse loss

+++ cnn-deblur/utils/custom_losses.py
@@ -28,7 +28,7 @@ def ms_mse(sharp_pyramid: List[tf.Tensor],
     loss = 0.
     for scale_trueY, scale_predY in zip(sharp_pyramid, predicted_pyramid):
         scale_shape = tf.shape(scale_trueY)[1:]
-        norm_factor = tf.reduce_prod(scale_shape)
+        norm_factor = tf.cast(tf.reduce_prod(scale_shape), dtype='float32')
         scale_loss = tf.reduce_sum(mse(scale_trueY, scale_predY)) / norm_factor
         loss += scale_loss
     return 1./(2. * num_scales) * loss
